Refactor: 암호화된 회원 캐싱 클래스 추가
- encryptor를 빈으로 등록하고 캐싱하면 모든 암호화 요청이 캐싱되기 때문에 회원 식별자만 캐싱하도록 별도의 클래스를 생성하였습니다.
