+++ src/main/java/com/flab/dduikka/config/CacheConfig.java
@@ -1,0 +1,16 @@
+package com.flab.dduikka.config;
+
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@EnableCaching
+@Configuration
+public class CacheConfig {
+	@Bean
+	public CacheManager cacheManager() {
+		return new ConcurrentMapCacheManager("MemberIdentifier");
+	}
+}
