+++ src/main/java/com/flab/dduikka/domain/livechat/application/LiveChatService.java
@@ -1,11 +1,10 @@
 package com.flab.dduikka.domain.livechat.application;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.flab.dduikka.common.util.SHA256Encryptor;
+import com.flab.dduikka.common.encryption.EncryptedMemberIdentifierCache;
 import com.flab.dduikka.common.validator.CustomValidator;
 import com.flab.dduikka.domain.livechat.domain.LiveChat;
 import com.flab.dduikka.domain.livechat.dto.LiveChatMessage;
@@ -21,14 +20,14 @@ import lombok.RequiredArgsConstructor;
 public class LiveChatService {
 
 	private final LiveChatRepository liveChatRepository;
-	private final SHA256Encryptor sha256Encryptor;
+	private final EncryptedMemberIdentifierCache cachedEncryptor;
 	private final CustomValidator validator;
 
 	public LiveChatResponse createMessage(long sessionId, LiveChatMessage message) {
 		LiveChat newLiveChat = LiveChatMessage.to(sessionId, message);
 		validator.validateObject(newLiveChat);
 		LiveChat createdLiveChat = liveChatRepository.addLiveChat(newLiveChat);
-		String encryptSessionId = sha256Encryptor.hashSHA256(String.valueOf(sessionId));
+		String encryptSessionId = cachedEncryptor.cacheEncryptedMemberIdentifier(String.valueOf(sessionId));
 		return LiveChatResponse.from(createdLiveChat, encryptSessionId);
 	}
 
@@ -38,9 +37,9 @@ public class LiveChatService {
 			liveChats.stream()
 				.map(liveChat -> LiveChatResponse.from(
 					liveChat,
-					sha256Encryptor.hashSHA256(String.valueOf(liveChat.getMemberId()))
+					cachedEncryptor.cacheEncryptedMemberIdentifier(String.valueOf(liveChat.getMemberId()))
 				))
-				.collect(Collectors.toList()));
+				.toList());
 	}
 
 	public void deleteLiveChat(long liveChatId, long memberId) {
